Reverted to working version

+++ src/app/components/app-users/app-users.component.ts
@@ -63,11 +63,11 @@ export class AppUsersComponent implements OnInit {
   }
 
   deleteUser(appUser:AppUser)
-  {
-    if(confirm("Delete user '"+appUser.name+" "+ appUser.surname+" ?")) {
-      this.appUsersService.deleteUser(appUser).subscribe(response=>{
-        this.loadGroup();
-      })
+   {
+     if(confirm("Delete user '"+appUser.name+" "+ appUser.surname+" and all payments and attendances connected?")) {
+       this.appUsersService.deleteUser(appUser).subscribe(response=>{
+         this.loadGroup();
+       })
     }
   }
 }
